+++ src/c_cmds.c
@@ -967,7 +967,7 @@ consolecmd_t consolecmds[] =
     CCMD(regenhealth, "", "", game_ccmd_func1, regenhealth_func2, true, "[" BOLD("on") "|" BOLD("off") "]",
         "Toggles regenerating your health by 1% every second when it's less than 100%."),
     CCMD(releasenotes, "", "", null_func1, releasenotes_func2, false, "",
-        "Opens the list of what's changed since the last version of " ITALICS(DOOMRETRO_NAME ".")),
+        "Opens a list of what's changed since the last version of " ITALICS(DOOMRETRO_NAME ".")),
     CCMD(remove, "", "", kill_func1, kill_func2, true, REMOVECMDFORMAT,
         "Removes all " BOLD("decorations") ", " BOLD("corpses") ", " BOLD("bloodsplats") ", " BOLD("items") ", certain "
         BOLDITALICS("items") ", or " BOLD("everything") "."),
@@ -7650,7 +7650,7 @@ static void regenhealth_func2(char *cmd, char *parms)
 //
 static void releasenotes_func2(char *cmd, char *parms)
 {
-    C_Output("Opening " ITALICS(DOOMRETRO_NAME "'s") " release notes...");
+    C_Output("Opening the release notes for " ITALICS(DOOMRETRO_NAMEANDVERSIONSTRING "..."));
 
 #if defined(_WIN32)
     if (!ShellExecute(NULL, "open", DOOMRETRO_RELEASENOTESURL, NULL, NULL, SW_SHOWNORMAL))
@@ -7659,7 +7659,7 @@ static void releasenotes_func2(char *cmd, char *parms)
 #elif defined(__APPLE__)
     if (!system("open " DOOMRETRO_RELEASENOTESURL))
 #endif
-        C_Warning(0, ITALICS(DOOMRETRO_NAME "'s") " release notes wouldn't open.");
+        C_Warning(0, "The release notes for " ITALICS(DOOMRETRO_NAMEANDVERSIONSTRING) " wouldn't open!");
 }
 
 //